+++ src/components/Oauth/NoWorkspaceEntry/NoWorkspaceOauthFlow.tsx
@@ -7,7 +7,7 @@ import { useCallback, useState } from 'react';
 import { useApiKey } from '../../../context/ApiKeyContextProvider';
 import { useProject } from '../../../context/ProjectContextProvider';
 import { capitalize } from '../../../utils';
-import { fetchOAuthPopupURL } from '../fetchOAuthCallbackURL';
+import { fetchOAuthPopupURL } from '../fetchOAuthPopupURL';
 import { OAuthWindow } from '../OAuthWindow/OAuthWindow';
 
 import { LandingContent } from './LandingContent';
